refactor: 들여쓰기 및 EmotionType 버튼 변수명 수정
