+++ src/components/common/buttons/Back/BtnBack.jsx
@@ -1,9 +1,9 @@
 import * as S from './BtnBack.style'
 
 export default function BtnBack({handleClick}){
-    return(
-        <S.ButtonWrapper onClick={()=>{handleClick()}}> 
-            <S.BackButton/>
-        </S.ButtonWrapper>
-    )
+	return(
+		<S.ButtonWrapper onClick={()=>{handleClick()}}> 
+			<S.BackButton/>
+		</S.ButtonWrapper>
+	)
 }
