+++ src/components/common/buttons/Back/BtnBack.style.js
@@ -2,9 +2,9 @@ import styled from "styled-components";
 import { IcBackButton } from "../../../../assets/svg";
 
 export const ButtonWrapper =styled.button`
-    display: flex;
-    background-color: transparent;
+	display: flex;
+	background-color: transparent;
 `
 export const BackButton = styled(IcBackButton)`
-    display: flex;
+	display: flex;
 `
