+++ src/components/common/buttons/complete/BtnComplete.jsx
@@ -1,9 +1,9 @@
 import * as S from './BtnComplete.style'
 
 export default function BtnComplete({onClick, children}){
-    return(
-        <S.ButtonWrapper onClick={onClick}>
-            {children}
-        </S.ButtonWrapper>
-    )
+	return(
+		<S.ButtonWrapper onClick={onClick}>
+			{children}
+		</S.ButtonWrapper>
+	)
 }
