+++ src/pages/SearchByEmotion/SearchByEmotionPopUp/SearchByEmotionPopUp.jsx
@@ -6,7 +6,7 @@ import { useNavigate } from "react-router-dom";
 import usePostFeelingList from "../../../hooks/queries/mypage/usePostFeelingList";
 import { useModal } from '../../../hooks/common/useModal';
 import DiaryErrorModal from '../../../components/Modal/DiaryErrorModal';
-import BtnEmotionType from '../../../components/common/buttons/EmotionType/EmotionType';
+import BtnEmotionType from '../../../components/common/buttons/EmotionType/BtnEmotionType';
 
 export default function SearchByEmotionPopUp(){
   const [diaryList, setDiaryList] = useRecoilState(diaryListAtom);
